System.UDev.Context

Description

The context contains the default values read from the udev config file, and is passed to all library operations.

Context

data UDev

Opaque object representing the library context.

class UDevChild a where

Family of udev resources.

Methods

getUDev :: a -> UDev

Get the context a resource belong to.

newUDev :: IO UDev

Create udev library context. This reads the udev configuration file, and fills in the default values.

withUDev :: (UDev -> IO a) -> IO a

Like newUDev but context will be released at exit.

Logging

data Priority

Log message priority.

Constructors

LogError
error conditions

LogInfo
informational

LogDebug
debug-level messages

type Logger

Arguments

= UDev

-> Priority
message priority;

-> ByteString
position: file

-> Int
position: line

-> ByteString
position: function

-> ByteString
message body

-> IO ()

Logger function will called by udev on events.

getLogPriority :: UDev -> IO Priority

The initial logging priority is read from the udev config file at startup.

setLogPriority :: UDev -> Priority -> IO ()

Set the current logging priority. The value controls which messages are logged.

setLogger :: UDev -> Logger -> IO ()

The built-in logging writes to stderr. It can be overridden by a custom function, to plug log messages into the users' logging functionality.

defaultLogger :: Logger

Default logger will just print %PRIO %FILE:%LINE:n%FN: %FORMAT to stdout.

User data

getUserdata :: UDev -> IO a

Retrieve stored data pointer from library context. This might be useful to access from callbacks like a custom logging function.

setUserdata :: UDev -> a -> IO ()

Store custom userdata in the library context.