System.UDev.Enumerate

Description

Lookup devices in the sys filesystem, filter devices by properties, and return a sorted list of devices.

Documentation

data Enumerate

Opaque object representing one device lookup/sort context.

newEnumerate :: UDev -> IO Enumerate

Create an enumeration context to scan /sys.

Match

type Subsystem = ByteString

Kernel subsystem string.

addMatchSubsystem

Arguments

:: Enumerate
context

-> Subsystem
filter for a subsystem of the device to include in the list

-> IO ()
can throw exception

Match only devices belonging to a certain kernel subsystem.

addNoMatchSubsystem

Arguments

:: Enumerate
context

-> Subsystem
filter for a subsystem of the device to exclude from the list

-> IO ()
can throw exception

Match only devices not belonging to a certain kernel subsystem.

type SysAttr = ByteString

/sys attribute string.

type SysValue = ByteString

Attribute specific /sys value string. Can be an int or identifier depending on attribute.

addMatchSysattr

Arguments

:: Enumerate
context

-> SysAttr
filter for a sys attribute at the device to include in the list

-> Maybe SysValue
optional value of the sys attribute

-> IO ()
can throw exception

Match only devices with a certain /sys device attribute.

addNoMatchSysattr

Arguments

:: Enumerate
context

-> ByteString
filter for a sys attribute at the device to exclude from the list

-> Maybe ByteString
optional value of the sys attribute

-> IO ()

Match only devices not having a certain /sys device attribute.

addMatchProperty

Arguments

:: Enumerate
context

-> ByteString
filter for a property of the device to include in the list

-> ByteString
value of the property

-> IO ()

Match only devices with a certain property.

addMatchTag

Arguments

:: Enumerate
context

-> ByteString
filter for a tag of the device to include in the list

-> IO ()

Match only devices with a certain tag.

addMatchParent

Arguments

:: Enumerate
context

-> Device
parent device where to start searching

-> IO ()
can throw exception

Return the devices on the subtree of one given device. The parent itself is included in the list.

A reference for the device is held until the udev_enumerate context is cleaned up.

addMatchIsInitialized :: Enumerate -> IO ()

Match only devices which udev has set up already.

addMatchSysname

Arguments

:: Enumerate
context

-> ByteString
filter for the name of the device to include in the list

-> IO ()
can throw exception

Match only devices with a given /sys device name.

addSyspath

Arguments

:: Enumerate
context

-> ByteString
path of a device

-> IO ()
can throw exception

Add a device to the list of devices, to retrieve it back sorted in dependency order.

Scan

scanDevices :: Enumerate -> IO ()

Scan /sys for all devices which match the given filters.

scanSubsystems :: Enumerate -> IO ()

Scan /sys for all devices which match the given filters.

Query

getListEntry :: Enumerate -> IO List

Get the first entry of the sorted list of device paths.