System.UDev.HWDB

Description

Retrieve properties from the hardware database.

Documentation

data HWDB

Opaque object representing the hardware database.

newHWDB :: UDev -> IO HWDB

Create a hardware database context to query properties for devices.

getPropertiesList :: HWDB -> ByteString -> IO List

Lookup a matching device in the hardware database. The lookup key is a modalias string, whose formats are defined for the Linux kernel modules. Examples are: pci:v00008086d00001C2D*, usb:v04F2pB221*. The first entry of a list of retrieved properties is returned.