System.UDev.List

Description

Libudev list operations.

Documentation

data List

Opaque object representing one entry in a list. An entry contains contains a name, and optionally a value.

getNext :: List -> IO (Maybe List)

Get the next entry from the list.

getByName :: List -> IO (Maybe List)

Lookup an entry in the list with a certain name.

getName :: List -> IO ByteString

Get the name of a list entry.

getValue :: List -> IO ByteString

Get the value of list entry.