System.UDev.Monitor

Description

Connects to a device event source.

Documentation

data Monitor

Opaque object handling an event source.

Creation

data SourceId

Event source identifier.

udevId :: SourceId

Events are sent out after udev has finished its event processing, all rules have been processed, and needed device nodes are created.

kernelId :: SourceId

Events are sent out just after kernel processes them.

Applications should usually not connect directly to the kernel events, because the devices might not be useable at that time, before udev has configured them, and created device nodes. Use kernelId instead.

newFromNetlink :: UDev -> SourceId -> IO Monitor

Create new udev monitor and connect to a specified event source.

Receiving

enableReceiving :: Monitor -> IO ()

Binds the udev_monitor socket to the event source.

setReceiveBufferSize :: Monitor -> Int -> IO ()

Set the size of the kernel socket buffer.

getFd :: Monitor -> IO Fd

Retrieve the socket file descriptor associated with the monitor.

getHandle :: Monitor -> IO Handle

Retrieve the socket handle associated with the monitor.

receiveDevice :: Monitor -> IO Device

Receive data from the udev monitor socket, allocate a new udev device, fill in the received data, and return the device.

Filter

filterAddMatchSubsystemDevtype :: Monitor -> ByteString -> ByteString -> IO ()

Filter events by subsystem and device type.

The filter must be installed before the monitor is switched to listening mode.

filterAddMatchTag :: Monitor -> ByteString -> IO ()

The filter must be installed before the monitor is switched to listening mode.

filterUpdate :: Monitor -> IO ()

Update the installed socket filter. This is only needed, if the filter was removed or changed.

filterRemove :: Monitor -> IO ()

Remove all filters from monitor.